Improve overflow_correction
In case we work on Altitude Radio we can use Fast phase to detect when
the aircraft was on the ground. In those points in time the Altitude
Radio should be 0. The pin_to_ground() subfunction does exactly what it
says on the tin: postprocesses the values for slices known to be on the
ground to be zero.