create_ktis_on_state_change allow slice start idx
KeyTimeInstanceNode.create_ktis_on_state_change will detect state
changes on the rounded index of the start of the slice. This accounts
for a 'gray area' between samples where the state of a multi-state
parameter is unknown.

As we are nudging the KTI index to be half a sample to account for the
unknown state of the inbetween samples, it is possible for the to be
nudged before the start of the section. This causes a discrepancy where
Gear Up can occur before Liftoff. In this case index will be nudged to
the start index of the section.